Moved /core/server/data/xml to /core/frontend/services
- Didn't move it into separate /data folder as it seems to better fit into 'services' category.
