+++ core/frontend/services/sitemap/handler.js
@@ -1,4 +1,4 @@
-const config = require('../../../config'),
+const config = require('../../../server/config'),
     Manager = require('./manager'),
     manager = new Manager();
