+++ core/frontend/services/sitemap/page-generator.js
