+++ core/frontend/services/sitemap/post-generator.js
