+++ core/frontend/services/sitemap/tag-generator.js
