+++ core/frontend/services/sitemap/user-generator.js
