+++ core/frontend/services/sitemap/utils.js
@@ -1,4 +1,4 @@
-var urlUtils = require('../../../lib/url-utils'),
+var urlUtils = require('../../../server/lib/url-utils'),
     sitemapsUtils;
 
 sitemapsUtils = {
