+++ core/server/web/site/app.js
@@ -8,8 +8,8 @@ const apps = require('../../services/apps');
 const constants = require('../../lib/constants');
 const storage = require('../../adapters/storage');
 const urlService = require('../../../frontend/services/url');
-const urlUtils = require('../../lib/url-utils');
-const sitemapHandler = require('../../data/xml/sitemap/handler');
+const urlUtils = require('../../../server/lib/url-utils');
+const sitemapHandler = require('../../../frontend/services/sitemap/handler');
 const themeMiddleware = require('../../../frontend/services/themes').middleware;
 const membersService = require('../../services/members');
 const siteRoutes = require('./routes');
